Fix some clippy lints for rust 1.66.

+++ src/tensor/convert.rs
@@ -7,7 +7,7 @@ use std::convert::{TryFrom, TryInto};
 impl<T: Element> From<&Tensor> for Vec<T> {
     fn from(tensor: &Tensor) -> Vec<T> {
         let numel = tensor.numel();
-        let mut vec = vec![T::ZERO; numel as usize];
+        let mut vec = vec![T::ZERO; numel];
         tensor.to_device(crate::Device::Cpu).to_kind(T::KIND).copy_data(&mut vec, numel);
         vec
     }
