+++ src/tensor/index.rs
@@ -305,11 +305,11 @@ impl Tensor {
                 ),
                 Narrow(Unbounded, Unbounded) => (curr_tensor, curr_idx + 1),
                 Narrow(Included(start), Unbounded) => {
-                    let dim_len = curr_tensor.size()[curr_idx as usize] as i64;
+                    let dim_len = curr_tensor.size()[curr_idx as usize];
                     (curr_tensor.narrow(curr_idx, *start, dim_len - *start), curr_idx + 1)
                 }
                 Narrow(Excluded(start), Unbounded) => {
-                    let dim_len = curr_tensor.size()[curr_idx as usize] as i64;
+                    let dim_len = curr_tensor.size()[curr_idx as usize];
                     (curr_tensor.narrow(curr_idx, *start + 1, dim_len - *start - 1), curr_idx + 1)
                 }
                 Narrow(Unbounded, Included(end)) => {
