+++ src/tensor/npy.rs
@@ -222,7 +222,7 @@ impl crate::Tensor {
         f.write_all(&[(header.len() % 256) as u8, (header.len() / 256) as u8])?;
         f.write_all(header.as_bytes())?;
         let numel = self.numel();
-        let mut content = vec![0u8; (numel * kind.elt_size_in_bytes()) as usize];
+        let mut content = vec![0u8; numel * kind.elt_size_in_bytes()];
         self.f_copy_data_u8(&mut content, numel)?;
         f.write_all(&content)?;
         Ok(())
